feat: moved ShippingLabelCarrierRowViewModel into a separate file
